Rename bitly class from Bitly to BitlyLinkShortener; Move bitly-specific interfaces into bitly class file
