+++ server/src/core/bootstrap.ts
@@ -12,7 +12,7 @@ import { EventBus } from './event';
 import { UserNotifications } from './user-notification';
 import { Statistics } from './stat';
 import { DistributionReports } from './distribution-report';
-import { Bitly, Links } from './link-generator';
+import { BitlyLinkShortener, Links } from './link-generator';
 
 export async function bootstrap(config: AppConfig): Promise<App> {
   const client = await getDbConnection(config.dbUri);
@@ -81,7 +81,7 @@ export async function bootstrap(config: AppConfig): Promise<App> {
 
   const stats = new Statistics(db);
 
-  const bitly = new Bitly({ apiKey: config.bitlyApiKey, apiLink: config.bitlyApiLink });
+  const bitly = new BitlyLinkShortener({ apiKey: config.bitlyApiKey, apiLink: config.bitlyApiLink });
 
   const links = new Links({ baseUrl: config.webappBaseUrl, bitly });
