+++ server/src/core/link-generator/bitly-link-shortener.ts
@@ -1,13 +1,44 @@
-import { BitlyService, ShortenLinkRes } from './types';
 import axios from 'axios';
 import { rethrowIfAppError, createBitlyApiError, createLinkShorteningFailedError } from '../error';
+
+export interface DeepLink {
+  guid: string,
+  bitlink: string,
+  app_uri_path: string,
+  install_url: string,
+  app_guid: string,
+  os: string,
+  install_type: string,
+  created: string,
+  modified: string,
+  brand_guid: string
+}
+
+export interface ShortenLinkRes {
+  references: any,
+  link: string,
+  id: string,
+  long_url: string,
+  title: string,
+  archived: boolean,
+  created_at: string,
+  created_by: string,
+  client_id: string,
+  custom_bitlinks: string[],
+  tags: string[],
+  deepLinks: DeepLink[]
+}
 
 export interface BitlyArgs {
   apiKey: string;
   apiLink: string;
 }
 
-export class Bitly implements BitlyService {
+export interface LinkShortener {
+  shortenLink(link: string): Promise<string>;
+}
+
+export class BitlyLinkShortener implements LinkShortener {
   private apiKey: string;
   private apiLink: string;
   private headers: any;
