+++ server/src/core/link-generator/index.ts
@@ -1,2 +1,2 @@
 export { Links } from './link-generator-service';
-export { Bitly } from './bitly-service';
+export { BitlyLinkShortener } from './bitly-link-shortener';
